+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 ### Release Notes
 
+### 2.0.0
+
+- Added support for specifying family shortcut global options, like `-4`.
+- Added support for batch mode.
+
 ### 2.0.0-beta.2
 
 - Added more commands: `addrlabel`, `neighbour`, `ntable`, `tunnel`, `tuntap`, `maddress`, `mroute`.
 - Add support for an array of tables to `routingTables.{add|del}` methods.
+- Added support for specifying family shortcut global options, like `-4`.
 
 ### 2.0.0-beta.1